Ajueste en la base de datos y conexion desde el codigo

+++ DB/ScriptDB.sql
@@ -1,19 +1,19 @@
 CREATE TABLE Faculty (
-    FacultyId INT NOT NULL,
+    FacultyId BIGSERIAL NOT NULL,
     FacultyName VARCHAR(255) NOT NULL,
     PRIMARY KEY (FacultyId)
 );
 
 CREATE TABLE Unity (
-    UnityId INT NOT NULL,
-    FacultyId INT NOT NULL,
+    UnityId BIGSERIAL NOT NULL,
+    FacultyId BIGSERIAL NOT NULL,
     UnityName VARCHAR(255) NOT NULL,
     PRIMARY KEY (UnityId),
     FOREIGN KEY (FacultyId) REFERENCES Faculty(FacultyId)
 );
 
 CREATE TABLE AcademicPeriod (
-    AcademicPeriodId INT NOT NULL,
+    AcademicPeriodId SERIAL NOT NULL,
     Description VARCHAR(255) NOT NULL,
     StartDate DATE NOT NULL,
     EndDate DATE NOT NULL,
@@ -21,22 +21,22 @@ CREATE TABLE AcademicPeriod (
 );
 
 CREATE TABLE AcademicProgram (
-    AcademicProgramId INT NOT NULL,
-    UnityId INT NOT NULL,
+    AcademicProgramId BIGSERIAL NOT NULL,
+    UnityId BIGSERIAL NOT NULL,
     AcademicProgramName VARCHAR(255) NOT NULL,
     PRIMARY KEY (AcademicProgramId),
     FOREIGN KEY (UnityId) REFERENCES Unity(UnityId)
 );
 
 CREATE TABLE Role (
-    RoleId INT NOT NULL,
+    RoleId BIGSERIAL NOT NULL,
     RoleName VARCHAR(255) NOT NULL,
     PRIMARY KEY (RoleId)
 );
 
 CREATE TABLE Users (
-    UserId INT NOT NULL,
-    RoleId INT NOT NULL,
+    UserId BIGSERIAL NOT NULL,
+    RoleId BIGSERIAL NOT NULL,
     FirstName VARCHAR(255) NOT NULL,
     LastName VARCHAR(255) NOT NULL,
     UserName VARCHAR(255) NOT NULL,
@@ -46,27 +46,27 @@ CREATE TABLE Users (
 );
 
 CREATE TABLE Signature (
-    SignatureId INT NOT NULL,
-    UnityId INT NOT NULL,
+    SignatureId BIGSERIAL NOT NULL,
+    UnityId BIGSERIAL NOT NULL,
     SignatureName VARCHAR(255) NOT NULL,
     PRIMARY KEY (SignatureId),
     FOREIGN KEY (UnityId) REFERENCES Unity(UnityId)
 );
 
 CREATE TABLE Instructor (
-    InstructorId INT NOT NULL,
-    UserId INT NOT NULL,
-    SignatureId INT NOT NULL,
+    InstructorId BIGSERIAL NOT NULL,
+    UserId BIGSERIAL NOT NULL,
+    SignatureId BIGSERIAL NOT NULL,
     PRIMARY KEY (InstructorId),
     FOREIGN KEY (UserId) REFERENCES Users(UserId),
     FOREIGN KEY (SignatureId) REFERENCES Signature(SignatureId)
 );
 
 CREATE TABLE SignatureGroup (
-    SignatureGroupId INT NOT NULL,
-    AcademicPeriodId INT NOT NULL,
-    InstructorId INT NOT NULL,
-    SignatureId INT NOT NULL,
+    SignatureGroupId BIGSERIAL NOT NULL,
+    AcademicPeriodId BIGSERIAL NOT NULL,
+    InstructorId BIGSERIAL NOT NULL,
+    SignatureId BIGSERIAL NOT NULL,
     AvailablePlaces INT NOT NULL,
     GroupName VARCHAR(255) NOT NULL,
     PRIMARY KEY (SignatureGroupId),
@@ -76,23 +76,23 @@ CREATE TABLE SignatureGroup (
 );
 
 CREATE TABLE SignatureProgram (
-    SignatureProgramId INT NOT NULL,
-    AcademicProgramId INT NOT NULL,
-    SignatureId INT NOT NULL,
+    SignatureProgramId BIGSERIAL NOT NULL,
+    AcademicProgramId BIGSERIAL NOT NULL,
+    SignatureId BIGSERIAL NOT NULL,
     PRIMARY KEY (SignatureProgramId),
     FOREIGN KEY (AcademicProgramId) REFERENCES AcademicProgram(AcademicProgramId),
     FOREIGN KEY (SignatureId) REFERENCES Signature(SignatureId)
 );
 
 CREATE TABLE Building (
-    BuildingId INT NOT NULL,
+    BuildingId BIGSERIAL NOT NULL,
     BuildingNumber INT NOT NULL,
     PRIMARY KEY (BuildingId)
 );
 
 CREATE TABLE Classroom (
-    ClassroomId INT NOT NULL,
-    BuildingId INT NOT NULL,
+    ClassroomId BIGSERIAL NOT NULL,
+    BuildingId BIGSERIAL NOT NULL,
     ClassroomNumber INT NOT NULL,
     PRIMARY KEY (ClassroomId),
     FOREIGN KEY (BuildingId) REFERENCES Building(BuildingId)
@@ -101,9 +101,9 @@ CREATE TABLE Classroom (
 
 
 CREATE TABLE SignatureGroupSchedule (
-    SignatureGroupScheduleId INT NOT NULL,
-    SignatureGroupId INT NOT NULL,
-    ClassroomId INT,
+    SignatureGroupScheduleId BIGSERIAL NOT NULL,
+    SignatureGroupId BIGSERIAL NOT NULL,
+    ClassroomId BIGSERIAL,
     StartHour VARCHAR(255) NOT NULL,
     EndHour VARCHAR(255) NOT NULL,
     ClassDay VARCHAR(255) NOT NULL,
@@ -113,18 +113,18 @@ CREATE TABLE SignatureGroupSchedule (
 );
 
 CREATE TABLE Studious (
-    StudiousId INT NOT NULL,
-    UserId INT NOT NULL,
-    AcademicProgramId INT NOT NULL,
+    StudiousId BIGSERIAL NOT NULL,
+    UserId BIGSERIAL NOT NULL,
+    AcademicProgramId BIGSERIAL NOT NULL,
     PRIMARY KEY (StudiousId),
     FOREIGN KEY (UserId) REFERENCES Users(UserId),
     FOREIGN KEY (AcademicProgramId) REFERENCES AcademicProgram(AcademicProgramId)
 );
 
 CREATE TABLE SignatureInscribed (
-    SignatureGroupAttendanceId INT NOT NULL,
-    StudiousId INT NOT NULL,
-    SignatureGroupId INT NOT NULL,
+    SignatureGroupAttendanceId BIGSERIAL NOT NULL,
+    StudiousId BIGSERIAL NOT NULL,
+    SignatureGroupId BIGSERIAL NOT NULL,
     PRIMARY KEY (SignatureGroupAttendanceId, StudiousId, SignatureGroupId),
     FOREIGN KEY (StudiousId) REFERENCES Studious(StudiousId),
     FOREIGN KEY (SignatureGroupId) REFERENCES SignatureGroup(SignatureGroupId)
